% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calc_something}
\alias{calc_something}
\title{Imagine this is a resource-intensive calculation}
\usage{
calc_something()
}
\value{
results as a dataframe
}
\description{
Also imagine that it is always correct and you don't need to test it
}
