% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_binary_outcome.R
\name{check_binary_outcome}
\alias{check_binary_outcome}
\title{Check that the outcome column is binary}
\usage{
check_binary_outcome(dataset, outcome_colname)
}
\arguments{
\item{dataset}{dataframe with a column containing outcomes}

\item{outcome_colname}{name of the column containing outcomes}
}
\description{
Check that the outcome column is binary
}
