% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing.R
\name{check_missing}
\alias{check_missing}
\title{Throw an error if \code{x} contains \code{NA}.}
\usage{
check_missing(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Returns nothing if \code{x} is good.
}
