% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_frac.R
\name{is_frac}
\alias{is_frac}
\title{Check whether the number is a fraction.}
\usage{
is_frac(frac)
}
\arguments{
\item{frac}{fraction (numeric)}
}
\value{
TRUE if \code{frac} is in \link{0, 1}, else \code{FALSE}
}
\description{
Check whether the number is a fraction.
}
