% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Imagine this is the main function of a complicated script}
\usage{
main(do_calc = FALSE)
}
\arguments{
\item{do_calc}{whether to run \code{calc_something_big()} or not (default: \code{FALSE})}
}
\value{
results as a named list
}
\description{
Imagine this is the main function of a complicated script
}
