% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohwhaley.R
\name{say}
\alias{say}
\title{Whale speak}
\usage{
say(what)
}
\arguments{
\item{what}{text string: what you want the whale to say.}
}
\value{
text string with a text whale and it saying a phrase.
}
\description{
This function is super fun and cute.
}
\examples{
say('Hello')
}
