% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rel_abun.R
\name{to_rel_abun}
\alias{to_rel_abun}
\title{Calculate relative abundances}
\usage{
to_rel_abun(dat)
}
\arguments{
\item{dat}{data.frame of absolute OTU abundances with columns as OTUs and rows as samples}
}
\value{
data.frame of relative OTU abundances
}
\description{
Calculate relative abundances
}
\examples{
to_rel_abun(data.frame(otu1 = 1:3, otu2 = 0:2))
}
