% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohwhaley.R, R/phrases.R
\docType{data}
\name{phrases}
\alias{phrases}
\title{Whale pun phrases used in say()}
\format{
An object of class \code{character} of length 9.

An object of class \code{character} of length 9.
}
\usage{
phrases

phrases
}
\description{
Whale pun phrases used in say()

Whale pun phrases used in say()
}
\keyword{internal}
