% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrases.R
\docType{data}
\name{phrases}
\alias{phrases}
\title{Sample phrases for function \code{say}}
\format{
A character vector of phrases
}
\source{
MQ Uni STAT1378 Tutorial Week 6
}
\usage{
phrases
}
\description{
A character vector of sample phrases for the function \code{say}
}
\keyword{datasets}
