% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohwhaley.R
\name{say}
\alias{say}
\title{summon a whale for a customisable pick up}
\usage{
say(what)
}
\arguments{
\item{what}{text string: what you want it to say, otherwise a random phrase}
}
\value{
text string with a quirky phrase
}
\description{
Summon a whale for a customisable pick me up.
}
\examples{
say("You're whale-come")
}
