% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say.R
\name{say}
\alias{say}
\title{Title}
\usage{
say(what)
}
\arguments{
\item{what;}{you can enter one of the set messages}
}
\value{
Test String with funny whale message
}
\description{
Function to make you feel better
}
\examples{
say()
}
