% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say.R
\name{say}
\alias{say}
\title{say}
\usage{
say(what)
}
\arguments{
\item{what}{whatever phrase you want to echo}
}
\value{
Whale shaped message
}
\description{
Summon a whale for a customisable pick-me-up
}
\examples{
say("You're whale-com")
}
