% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohwhaley.R
\name{say}
\alias{say}
\title{Summon a whale for a customisable pick-me-up!}
\usage{
say(what)
}
\arguments{
\item{what}{Whatever phrase you want the whale to echo. If no phrase is chosen, a random one will be provided.}
}
\value{
Whale-shaped message
}
\description{
Summon a whale for a customisable pick-me-up!
}
\examples{
say("Great job!")
}
