% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohwhaley.R
\name{say}
\alias{say}
\title{Summon a whale for a customisable pick-me-up}
\usage{
say(what)
}
\arguments{
\item{what}{text string; This is the phrase you want the whale to echo. If not supplied a random phrase is chosen.}
}
\value{
text string; Whale shaped message
}
\description{
Summon a whale for a customisable pick-me-up
}
\examples{
say("You're whale-come")
}
