% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohwhaley.R
\name{say}
\alias{say}
\title{Deliver praise}
\usage{
say(what)
}
\arguments{
\item{what}{whatever phrase you want to be able to echo. if not supplied, a random phrase will be chosen.}
}
\value{
Whale shaped message
}
\description{
Summon a whale for a customisable pick me up
}
\examples{
say("You're whale-come")
}
