% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whale.R
\name{say}
\alias{say}
\title{Ohwhaley help}
\usage{
say(what)
}
\arguments{
\item{what}{text string; This is what you want the whale to say. If not supplied, a random phrase is chosen.}
}
\value{
Text string
}
\description{
You are talking to a whale. Think about that.
}
\examples{
say("Whale-come to ohwhaley!")
}
