% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say.R
\name{say}
\alias{say}
\title{Summon a whale for customisable pick-me-up}
\usage{
say(what)
}
\arguments{
\item{what}{Whatever phrase you want the whale to echo. If not supplied, random message outputed}
}
\value{
whale shaped message
}
\description{
Summon a whale for customisable pick-me-up
}
\examples{
say("You're whalecome")
}
