% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say.R
\name{say}
\alias{say}
\title{Summon a whale for a customisable pick-me-up}
\usage{
say(what)
}
\arguments{
\item{what}{Whatever phrase you want whale to echo. If not supplied, a random phrase is chosen}
}
\value{
Whale shaped message
}
\description{
Summon a whale for a customisable pick-me-up
}
\examples{
say("You're whale-come")
}
