% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say.R
\name{say}
\alias{say}
\title{Pick-me-up whale}
\usage{
say(what)
}
\arguments{
\item{what}{text string; pick me up message, if blank randomly chosen}
}
\value{
a cute whale wholesome message
}
\description{
Pick-me-up whale
}
\examples{
say("You're whale-come!")
}
