% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohwhaley.R
\name{say}
\alias{say}
\title{Summon a whale for a customisable pik-me-up}
\usage{
say(what)
}
\arguments{
\item{what}{whatever phrase you want a whale to echo. if not supplied, a random phrase is chosen}
}
\value{
whale shaped message
}
\description{
Summon a whale for a customisable pik-me-up
}
\examples{
say("you're whale-come")

}
