% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say.R
\name{say}
\alias{say}
\title{Say whale puns or message}
\usage{
say(what)
}
\arguments{
\item{what}{character string to be printed, if NULL defaults to random whale pun}
}
\value{
Whale shapped message
}
\description{
Say whale puns or message
}
\examples{
say("Hi folks!")

}
