% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohwhaley.R
\name{say}
\alias{say}
\title{The Whale Message Function}
\usage{
say(what)
}
\arguments{
\item{what}{string; This is whatever you would like the whale to say.}
}
\value{
A whale with a message.
}
\description{
Summon a whale for a customisable pick-me-up message.
}
\examples{
say("You're very whale-come.")
say("You got this!")
say("Just keep swimming, swimming, swimming...")
}
